Move gitsigns to separate config file
